+++ Assets/Scripts/Player/HandleCombo.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class HandleCombo : StateMachineBehaviour
 {
-    [SerializeField] private float newDamageModifier;
+    [SerializeField] private int newDamageModifier;
     private static readonly int CanDoCombo = Animator.StringToHash("canDoCombo");
     // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
